Rename truffle-box to @truffle/box
