+++ packages/box/lib/config.js
