+++ packages/truffle-core/lib/commands/unbox.js
@@ -111,7 +111,7 @@ const command = {
   },
   run(options, done) {
     const Config = require("truffle-config");
-    const Box = require("truffle-box");
+    const Box = require("@truffle/box");
     const OS = require("os");
 
     const config = Config.default().with({
