+++ packages/truffle-core/package.json
@@ -44,7 +44,7 @@
     "spawn-args": "^0.1.0",
     "temp": "^0.8.3",
     "truffle-artifactor": "^4.0.29",
-    "truffle-box": "^1.0.35",
+    "@truffle/box": "^1.0.35",
     "truffle-compile": "^4.2.2",
     "truffle-config": "^1.1.19",
     "truffle-contract": "^4.0.30",
