+++ packages/truffle-core/test/config.js
@@ -1,7 +1,7 @@
 var assert = require("chai").assert;
 var fs = require("fs-extra");
 var glob = require("glob");
-var Box = require("truffle-box");
+var Box = require("@truffle/box");
 var Contracts = require("truffle-workflow-compile");
 var Ganache = require("ganache-core");
 var provision = require("truffle-provisioner");
