+++ packages/truffle-core/test/lib/commands/compile.js
@@ -1,5 +1,5 @@
 var assert = require("chai").assert;
-var Box = require("truffle-box");
+var Box = require("@truffle/box");
 var Contracts = require("truffle-workflow-compile");
 var Artifactor = require("truffle-artifactor");
 var Resolver = require("truffle-resolver");
