+++ packages/truffle-core/test/lib/create.js
@@ -2,7 +2,7 @@ var assert = require("chai").assert;
 var path = require("path");
 var fs = require("fs-extra");
 var glob = require("glob");
-var Box = require("truffle-box");
+var Box = require("@truffle/box");
 var Create = require("../../lib/create");
 var dir = require("node-dir");
 var Resolver = require("truffle-resolver");
