+++ packages/truffle-core/test/plugin.js
@@ -105,12 +105,12 @@ describe("plugin loader", () => {
     it("returns array of locally or globally installed options.plugins", () => {
       assert(
         pluginLoader.checkPluginModules({
-          plugins: ["truffle-box"],
+          plugins: ["@truffle/box"],
           working_directory: process.cwd()
         })
       );
       let pluginArray = pluginLoader.checkPluginModules({
-        plugins: ["truffle-box"],
+        plugins: ["@truffle/box"],
         working_directory: process.cwd()
       });
       assert(pluginArray);
@@ -118,12 +118,12 @@ describe("plugin loader", () => {
 
       assert(
         pluginLoader.checkPluginModules({
-          plugins: ["truffle-box", "truffle-config"],
+          plugins: ["@truffle/box", "truffle-config"],
           working_directory: process.cwd()
         })
       );
       pluginArray = pluginLoader.checkPluginModules({
-        plugins: ["truffle-box", "truffle-config"],
+        plugins: ["@truffle/box", "truffle-config"],
         working_directory: process.cwd()
       });
       assert(pluginArray);
@@ -135,14 +135,14 @@ describe("plugin loader", () => {
     it("throws when plugins are installed without a truffle-plugin.json configuration file", () => {
       assert.throws(
         () => {
-          pluginLoader.loadPluginModules(["truffle-box"]);
+          pluginLoader.loadPluginModules(["@truffle/box"]);
         },
         TruffleError,
         "TruffleError not thrown!"
       );
       assert.throws(
         () => {
-          pluginLoader.loadPluginModules(["truffle-box", "truffle-config"]);
+          pluginLoader.loadPluginModules(["@truffle/box", "truffle-config"]);
         },
         TruffleError,
         "TruffleError not thrown!"
@@ -243,14 +243,14 @@ describe("plugin loader", () => {
       it("throws when plugins are installed without a truffle-plugin.json configuration file", () => {
         assert.throws(
           () => {
-            pluginLoader.loadPluginModules(["truffle-box"]);
+            pluginLoader.loadPluginModules(["@truffle/box"]);
           },
           TruffleError,
           "TruffleError not thrown!"
         );
         assert.throws(
           () => {
-            pluginLoader.loadPluginModules(["truffle-box", "truffle-config"]);
+            pluginLoader.loadPluginModules(["@truffle/box", "truffle-config"]);
           },
           TruffleError,
           "TruffleError not thrown!"
