+++ packages/truffle-core/test/profiler.js
@@ -1,7 +1,7 @@
 var assert = require("chai").assert;
 var fs = require("fs-extra");
 var glob = require("glob");
-var Box = require("truffle-box");
+var Box = require("@truffle/box");
 var Profiler = require("truffle-compile/profiler");
 var Resolver = require("truffle-resolver");
 var Artifactor = require("truffle-artifactor");
