+++ packages/truffle-debugger/package.json
@@ -61,7 +61,7 @@
     "nyc": "^13.0.1",
     "remotedev-server": "^0.3.1",
     "truffle-artifactor": "^4.0.29",
-    "truffle-box": "^1.0.35",
+    "@truffle/box": "^1.0.35",
     "truffle-debug-utils": "^1.0.18",
     "truffle-migrate": "^3.0.31",
     "truffle-resolver": "^5.0.15",
