+++ packages/truffle-debugger/test/helpers.js
@@ -8,7 +8,7 @@ import Debug from "truffle-debug-utils";
 import Artifactor from "truffle-artifactor";
 import Web3 from "web3";
 import Migrate from "truffle-migrate";
-import Box from "truffle-box";
+import Box from "@truffle/box";
 import Resolver from "truffle-resolver";
 
 export async function prepareContracts(provider, sources = {}, migrations) {
