+++ packages/truffle/package.json
@@ -48,7 +48,7 @@
     "shebang-loader": "0.0.1",
     "stream-buffers": "^3.0.1",
     "tmp": "0.0.33",
-    "truffle-box": "^1.0.35",
+    "@truffle/box": "^1.0.35",
     "truffle-contract": "^4.0.30",
     "truffle-core": "^5.0.33",
     "truffle-debugger": "^5.0.25",
